+++ app/src/main/java/com/nickoehler/brawlhalla/clans/domain/ClanDataSource.kt
@@ -5,6 +5,6 @@ import com.nickoehler.brawlhalla.core.domain.util.Result
 
 interface ClanDataSource {
 
-    suspend fun getClan(clanId: Int): Result<ClanDetail, NetworkError>
+    suspend fun getClan(clanId: Long): Result<ClanDetail, NetworkError>
 
 }
