+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/domain/StatClan.kt
@@ -2,7 +2,7 @@ package com.nickoehler.brawlhalla.ranking.domain
 
 data class StatClan(
     val clanName: String,
-    val clanId: Int,
-    val clanXp: Int,
-    val personalXp: Int,
+    val clanId: Long,
+    val clanXp: Long,
+    val personalXp: Long,
 )
